zzw modifications
- adapt for spconv 2.X seires
- adapt for pytorch_lightning 2.0.8 and lightning.pytorch new api

+++ cont_assoc/utils/eval_np.py
@@ -153,7 +153,7 @@ class PanopticEval:
       pred_areas = np.array([counts_pred[id2idx_pred[id]] for id in pred_labels])
       intersections = counts_combo
       unions = gt_areas + pred_areas - intersections
-      ious = intersections.astype(np.float) / unions.astype(np.float)
+      ious = intersections.astype(np.float32) / unions.astype(np.float32)
 
 
       tp_indexes = ious > 0.5
